requirements.txt
streamlit
ultralytics
opencv-python
pillow
pandas
mysql-connector-python
